Tets22 move to Page Objects structure
#15
