+++ src/main/java/net/commments/ft/Test4.java
@@ -1,5 +1,6 @@
-package net.commments;
+package net.commments.ft;
 
+import org.hamcrest.MatcherAssert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -7,30 +8,32 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import static org.testng.Assert.assertEquals;
+//import org.openqa.selenium.firefox.FirefoxDriver;
 
-public class Test22 {
+public class Test4 {
 
     private final WebDriver[] driver = new WebDriver[1];
 
     @Test
     public void test() {
         this.driver().get("http://commentssprintone.azurewebsites.net");
-        this.driver().findElements(By.id("newbutton")).get(0).click();
-        this.driver().findElement(By.id("Text")).sendKeys("@#$%^&*&*");
-        this.driver().findElements(By.xpath("//*[@id=\"editor-navigation\"]/input[1]")).get(0).click();
-        String errorText = this.driver().findElement(By.id("errorfield")).getText();
-        assertEquals(errorText, "The Comment Text field should contain alphanumeric characters only");
+        this.driver().findElements(By.name("SelectedId")).get(0).click();
+        this.driver().findElement(By.xpath("//*[@value=\"Delete\"]")).click();
+        this.driver().findElement(By.xpath("//span[text()=\"Yes\"]")).click();
+        final String page = this.driver().getPageSource();
+        MatcherAssert.assertThat("Comment Text 0 is present", !page.contains("Comment Text 0"));
     }
 
     @BeforeMethod
     public void createDriver() {
+//     driver[0] = new FirefoxDriver();
         driver[0] = new ChromeDriver();
+
     }
 
     @AfterMethod
     public void closeDriver() {
-        driver[0].quit();
+        driver[0].close();
     }
 
     private WebDriver driver() {
