+++ src/main/java/net/commments/object/CommentFilter.java
@@ -1,5 +1,4 @@
 package net.commments.object;
 
 public interface CommentFilter {
-    CommentTable
 }
